Create by bhagyashri

+++ src/main/java/com/example/ATCProject/Repository/ApplicationRepo.java
@@ -1,8 +1,8 @@
 package com.example.ATCProject.Repository;
 
-import com.example.ATCProject.model.Application;
+import com.example.ATCProject.Entity.Application;
+import com.example.ATCProject.Entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationRepo extends JpaRepository<Application,Long> {
-
+public interface ApplicationRepo extends JpaRepository<Application,Integer> {
 }
