+++ src/main/java/com/example/ATCProject/Repository/CollageRepo.java
@@ -1,9 +1,8 @@
 package com.example.ATCProject.Repository;
 
 
-import com.example.ATCProject.DTO.CollegeDTO;
-import com.example.ATCProject.model.College;
+import com.example.ATCProject.Entity.College;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CollageRepo extends JpaRepository<College,Long> {
+public interface CollageRepo extends JpaRepository<College,Integer> {
 }
