+++ src/main/java/com/example/ATCProject/Repository/CompanyRepo.java
@@ -1,8 +1,9 @@
 package com.example.ATCProject.Repository;
 
-import com.example.ATCProject.model.Company;
+import com.example.ATCProject.Entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepo extends JpaRepository <Company,Long>{
+public interface CompanyRepo extends JpaRepository <Company,Integer>{
 
+   // CompanyDTO  Email();
 }
